update urban growth example to use cells as agents
